requirements.txt
absl-py==2.2.2
ale-py==0.8.1
annotated-types==0.7.0
anyio==4.5.2
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
cachetools==5.5.2
carla==0.9.11
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
grpcio==1.70.0
gym==0.23.0
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.16.0
h5py==3.11.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax-jumpy==1.0.0
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
open-clip-torch==2.20.0
openai==1.76.0
opencv-python==4.7.0.68
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==3.20.3
psutil==7.0.0
py-trees==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.1.4
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
python-box==7.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
shapely==2.0.7
shiboken6==6.6.3.1
Shimmy==0.2.1
six==1.17.0
sniffio==1.3.1
srunner==0.15
stable-baselines3==2.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.15
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
zipp==3.20.2
